Commit again sum.c, sum.h into sum/calculator-class

+++ sum.h
@@ -11,7 +11,7 @@
  * @param b 
  * @return int 
  */
-int sum(int a, int b);
+uint8_t sum(uint8_t a, uint8_t b);
 
 #endif
 /*******************************************************************************
